+++ packages/build-app-helpers/src/getRoutesByAppJson.ts
@@ -3,7 +3,8 @@ import { IRoute } from './types';
 
 interface IStaticConfig {
   routes: IRoute[];
-  window: any;
+  window?: any;
+  tabBar?: any;
 }
 
 interface IOptions {
@@ -24,11 +25,20 @@ export default function (target, { appJsonPath, appJsonContent }: IOptions) {
     throw new Error('routes should be an array in app.json.');
   }
 
-  return staticConfig.routes.filter((route) => {
+  const routes = staticConfig.routes.filter((route) => {
     if (Array.isArray(route.targets) && !route.targets.includes(target)) {
       return false;
     }
 
     return true;
   });
+
+  if (staticConfig?.tabBar?.source) {
+    routes.push({
+      ...staticConfig.tabBar,
+      __tabBar: true,
+    });
+  }
+
+  return routes;
 }