+++ packages/build-user-config/src/applyCliOption.js
@@ -1,3 +1,4 @@
+const path = require('path');
 const optionConfig = require('./config/option.config');
 
 module.exports = (api, options = {}) => {
@@ -19,7 +20,7 @@ module.exports = (api, options = {}) => {
     if (module !== false) {
       try {
         // eslint-disable-next-line
-        optionDefination.configWebpack = require(`./cliOption/${moduleName}`);
+        optionDefination.configWebpack = require(path.isAbsolute(moduleName) ? moduleName : `./cliOption/${moduleName}`);
       } catch (err) {
         log.error(err);
       }