+++ packages/plugin-webpack5/src/index.ts
@@ -2,17 +2,24 @@ import * as path from 'path';
 import * as fse from 'fs-extra';
 import { IPlugin, Json } from '@alib/build-scripts';
 import analyzeNext from './analyzeNext';
-import filterPackages from './filterPackages';
+import filterPackages, { IFilterOptions, IRule } from './filterPackages';
 import remoteConfig from './remoteConfig';
 import compileRemote from './compileRemote';
 
+interface IRemoteOptions extends IFilterOptions {
+  activeInBuild: boolean;
+  include?: IRule;
+  exclude?: IRule;
+  autoDetect?: boolean;
+}
 interface IOptions {
-  remoteRuntime?: boolean;
+  remoteRuntime?: boolean | IRemoteOptions;
+  bootstrap?: string;
 }
 
-const plugin: IPlugin = (api, options = {}) => {
+const plugin: IPlugin = async (api, options = {}) => {
   const { onGetWebpackConfig, registerUserConfig, context } = api;
-  const { remoteRuntime } = options as IOptions;
+  const { remoteRuntime, bootstrap } = options as IOptions;
   const { pkg, userConfig, webpack, command } = context;
   const { ModuleFederationPlugin } = (webpack as any).container;
 
@@ -36,23 +43,38 @@ const plugin: IPlugin = (api, options = {}) => {
     react: 'React',
     'react-dom': 'ReactDOM',
   };
-  // check @alifd/next
-  const [cssPath, removePackage] = analyzeNext(userConfig, context.rootDir);
-  if (removePackage) {
-    // compile next
-    delete pkgDeps[removePackage];
+  const externals = [];
+  if (userConfig?.externals) {
+    externals.push(userConfig.externals);
   }
-  // read pkgDep from cache
-  let cacheContent = {};
-  try {
-    cacheContent = fse.readJSONSync(depsPath);
-  } catch(err) {
-    // ignore err
+  if (!(userConfig?.externals as Json)?.react) {
+    externals.push(externalMap);
   }
   // filter dependencies
-  const compileKeys = filterPackages(Object.keys(pkgDeps));
-  const needCompile = activeRemoteRuntime && JSON.stringify(pkgDeps) !== JSON.stringify(cacheContent);
+  let compileKeys = await filterPackages(Object.keys(pkgDeps), context.rootDir, typeof remoteRuntime !== 'boolean' ? remoteRuntime : {});
+  let needCompile = false;
+
   if (activeRemoteRuntime) {
+    let cssPath = '';
+    // read pkgDep from cache
+    let cacheContent = {};
+    try {
+      cacheContent = fse.readJSONSync(depsPath);
+    } catch(err) {
+      // ignore err
+    }
+    if (pkgDeps['@alifd/next']) {
+      // check @alifd/next
+      const [nextCSS, removePackage] = analyzeNext(userConfig, context.rootDir);
+      cssPath = nextCSS;
+      if (removePackage) {
+        // compile next
+        delete pkgDeps[removePackage];
+        compileKeys = compileKeys.filter(compileKey => compileKey !== removePackage);
+      }
+    }
+    // check deps after remote package
+    needCompile = activeRemoteRuntime && JSON.stringify(compileKeys) !== JSON.stringify(cacheContent);
     // ensure folder before compile and copy
     fse.ensureDirSync(runtimeFolder);
     const externalBundles = [
@@ -66,11 +88,11 @@ const plugin: IPlugin = (api, options = {}) => {
       fse.copyFileSync(filePath, path.join(runtimeFolder, fileName));
       injectBundles.push(`/remoteRuntime/${fileName}`);
     });
-    remoteConfig(api, { remoteName, runtimeFolder, injectBundles, externalMap, compileKeys });
+    remoteConfig(api, { remoteName, runtimeFolder, injectBundles, externals, compileKeys, bootstrap });
   }
   // if missmatch cache compile remote runtime
   if (needCompile) {
-    compileRemote(api, { runtimeFolder, cacheFolder, externalMap, remoteEntry, remoteName, depsPath, compileKeys, pkgDeps });
+    compileRemote(api, { runtimeFolder, cacheFolder, externals, remoteEntry, remoteName, depsPath, compileKeys });
   }
 
   registerUserConfig({