+++ packages/plugin-webpack5/src/remoteConfig.ts
@@ -3,17 +3,21 @@ import * as fse from 'fs-extra';
 import * as cheerio from 'cheerio';
 import { IPluginAPI } from '@alib/build-scripts';
 
-export default (api: IPluginAPI, { remoteName, compileKeys, runtimeFolder, injectBundles, externalMap }) => {
-  const { context, getValue, modifyUserConfig, onGetWebpackConfig } = api;
-  const { userConfig } = context;
+export default (api: IPluginAPI, { remoteName, compileKeys, runtimeFolder, injectBundles, externals, bootstrap }) => {
+  const { getValue, modifyUserConfig, onGetWebpackConfig, context } = api;
   // create boostrap for mf
-  const bootstrapPath = path.join(getValue('TEMP_PATH'), 'bootstrap.ts');
-  fse.writeFileSync(bootstrapPath, 'import(\'../src/app\')', 'utf-8');
+  let bootstrapEntry = '';
+  if (!bootstrap) {
+    bootstrapEntry = path.join(getValue('TEMP_PATH'), 'bootstrap.ts');
+    fse.writeFileSync(bootstrapEntry, 'import(\'../src/app\')', 'utf-8');
+  } else {
+    bootstrapEntry = path.isAbsolute(bootstrap) ? bootstrap : path.join(context.rootDir, bootstrap);
+  }
   modifyUserConfig((modfiyConfig) => {
     const remotePlugins = [[require.resolve('./babelPluginRemote'), { libs: compileKeys, remoteName }]];
     return {
       babelPlugins: Array.isArray(modfiyConfig.babelPlugins) ? modfiyConfig.babelPlugins.concat(remotePlugins) : remotePlugins,
-      entry: bootstrapPath,
+      entry: bootstrapEntry,
       moduleFederation: {
         name: 'app',
         remoteType: 'window',
@@ -28,11 +32,6 @@ export default (api: IPluginAPI, { remoteName, compileKeys, runtimeFolder, injec
       return [[...args, { from: runtimeFolder, to: path.join(args[0].to, 'remoteRuntime') }]];
     });
 
-    const externals = [];
-    if (userConfig.externals) {
-      externals.push(userConfig.externals);
-    }
-    externals.push(externalMap);
     config.externals(externals);
     // inject runtime entry and externals umd
     if (config.plugins.get('HtmlWebpackPlugin')) {