+++ packages/plugin-webpack5/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.2
+
+- [feat] auto detect package for remote runtime
+
+## 1.2.1
+
+- [fix] active rule of remote runtime
+
 ## 1.2.0
 
 - [feat] support remote runtime by module federation